requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
tzdata==2023.3
psycopg2
Pillow
djangorestframework
django-filter
djangorestframework-gis
django-leaflet
python-dotenv
